Add book models and script to populate books

+++ book_library_app/models.py
@@ -102,7 +102,7 @@ class Book(db.Model):
     __tablename__ = 'books'
     id = db.Column(db.Integer, primary_key=True)
     title = db.Column(db.String(50), nullable=False)
-    isbn = db.Column(db.Integer, nullable=False, unique=True)
+    isbn = db.Column(db.BigInteger, nullable=False, unique=True)
     number_of_pages = db.Column(db.Integer, nullable=False)
     description = db.Column(db.Text)
     author_id = db.Column(db.Integer, db.ForeignKey('authors.id'), nullable=False)
